feat: palette 타입 지정 수정
