+++ src/components/topBar/backMenuTopBar/index.tsx
@@ -1,7 +1,7 @@
 import SvgIcons from '@/assets/svgIcons';
 import Typography from '@/components/typography';
+import palette from '@/styles/theme/color';
 
-import palette from '@/styles/colors';
 import {CommonTextProps} from '@/types/ui';
 import {StyleSheet, View} from 'react-native';
