+++ src/components/topBar/subTitleTopBar/index.tsx
@@ -1,5 +1,6 @@
 import Typography from '@/components/typography';
-import palette from '@/styles/colors';
+import palette from '@/styles/theme/color';
+
 import {CommonTextProps} from '@/types/ui';
 import {StyleSheet, View} from 'react-native';
