--- src/styles/colors/another.ts
@@ -1,6 +1,0 @@
-const another = {
-  Black: '#000000',
-  White: '#FFFFFF',
-};
-
-export default another;
