--- src/styles/colors/background.ts
@@ -1,6 +1,0 @@
-const background = {
-  Primary: '#F5F5F5',
-  Secondary: '#FFFFFF',
-};
-
-export default background;
