--- src/styles/colors/fill.ts
@@ -1,8 +1,0 @@
-const fill = {
-  Normal: '#EBEBEA',
-  Strong: '#DAD9D8',
-  Assistive: '#F8F8F7',
-  Disable: '#D1D1D1',
-};
-
-export default fill;
