--- src/styles/colors/line.ts
@@ -1,7 +1,0 @@
-const line = {
-  Normal: '#E6E6E5',
-  Assistive: '#F3F2F2',
-  Disable: '#D1D1D1',
-};
-
-export default line;
