--- src/styles/colors/primary.ts
@@ -1,10 +1,0 @@
-const primary = {
-  Dark: '#804000',
-  Deep: '#E69200',
-  Normal: '#FFCC00',
-  Alternative: '#FFF08F',
-  Assistive: '#FFFACC',
-  Light: '#FFFCE6',
-};
-
-export default primary;
