--- src/styles/colors/state.ts
@@ -1,5 +1,0 @@
-const state = {
-  Point: '#FF3344',
-};
-
-export default state;
