--- src/styles/colors/text.ts
@@ -1,9 +1,0 @@
-const text = {
-  Strong: '#000000',
-  Normal: '#201F1E',
-  Alternative: '#757470',
-  Assistive: '#C3C3C1',
-  Disable: '#E0E0E0',
-} as const;
-
-export default text;
