+++ src/styles/theme/color.ts
@@ -1,0 +1,41 @@
+const palette = {
+  Another: {
+    Black: '#000000',
+    White: '#FFFFFF',
+  },
+  Background: {
+    Primary: '#F5F5F5',
+    Secondary: '#FFFFFF',
+  },
+  Fill: {
+    Normal: '#EBEBEA',
+    Strong: '#DAD9D8',
+    Assistive: '#F8F8F7',
+    Disable: '#D1D1D1',
+  },
+  Line: {
+    Normal: '#E6E6E5',
+    Assistive: '#F3F2F2',
+    Disable: '#D1D1D1',
+  },
+  Primary: {
+    Dark: '#804000',
+    Deep: '#E69200',
+    Normal: '#FFCC00',
+    Alternative: '#FFF08F',
+    Assistive: '#FFFACC',
+    Light: '#FFFCE6',
+  },
+  State: {
+    Point: '#FF3344',
+  },
+  Text: {
+    Strong: '#000000',
+    Normal: '#201F1E',
+    Alternative: '#757470',
+    Assistive: '#C3C3C1',
+    Disable: '#E0E0E0',
+  },
+} as const;
+
+export default palette;
